+++ extensions/amp-story/1.0/amp-story-page-attachment.js
@@ -18,9 +18,8 @@ import {Action, StateProperty} from './amp-story-store-service';
 import {DraggableDrawer, DrawerState} from './amp-story-draggable-drawer';
 import {HistoryState, setHistoryState} from './utils';
 import {Services} from '../../../src/services';
-import {StoryAnalyticsEvent} from '../../../src/analytics';
+import {StoryAnalyticsEvent, getAnalyticsService} from './story-analytics';
 import {dev} from '../../../src/log';
-import {getAnalyticsService} from './story-analytics';
 import {getState} from '../../../src/history';
 import {htmlFor} from '../../../src/static-template';
 import {toggle} from '../../../src/style';
